Move particle rewriter code into combined ParticleRewriter
Removes all the particle special handling in item/entity rewriters and adds a global ParticleRewriter interface to the api.
