+++ common/src/main/java/com/viaversion/viaversion/protocols/v1_19_4to1_20/rewriter/ItemPacketRewriter1_20.java
@@ -59,7 +59,6 @@ public final class ItemPacketRewriter1_20 extends ItemRewriter<ClientboundPacket
         registerMerchantOffers1_19(ClientboundPackets1_19_4.MERCHANT_OFFERS);
         registerSetCreativeModeSlot(ServerboundPackets1_19_4.SET_CREATIVE_MODE_SLOT);
         registerContainerSetData(ClientboundPackets1_19_4.CONTAINER_SET_DATA);
-        registerLevelParticles1_19(ClientboundPackets1_19_4.LEVEL_PARTICLES);
 
         protocol.registerClientbound(ClientboundPackets1_19_4.UPDATE_ADVANCEMENTS, wrapper -> {
             wrapper.passthrough(Types.BOOLEAN); // Reset/clear
